% Generated by roxygen2 (4.0.1): do not edit by hand
\name{admb.pairs}
\alias{admb.pairs}
\title{Examine MCMC output from an ADMB model.}
\usage{
admb.pairs(posterior, diag = c("acf", "hist", "trace"), fits = NULL,
  acf.ylim = c(-1, 1), ymult = NULL, axis.col = gray(0.5),
  limits = NULL, ...)
}
\arguments{
\item{posterior}{Dataframe containing the MCMC output, as read in using
function \link{\code{run.mcmc}}}

\item{diag}{What type of plot to include on the diagonal, options are
'acf' which plots the autocorrelation function \code{acf}, 'hist' shows
marginal posterior histograms, and 'trace' the trace plot.}

\item{fits}{A list containing the fitted information, read in using xxx}

\item{acf.ylim}{If using the acf function on the diagonal, specify the y
limit. The default is c(-1,1).}

\item{ymult}{A vector of length ncol(posterior) specifying how much room to
give when using the hist option for the diagonal. For use if the label
is blocking part of the plot.}

\item{limits}{A list containing the ranges for each parameter to use in
plotting.}
}
\description{
This function is useful for checking covergence and posterior
properties.
}

