% Generated by roxygen2 (4.0.1): do not edit by hand
\name{boundp}
\alias{boundp}
\title{Bounding transformation used by ADMB to map unbounded (-Infinity,
Infinity) to (minb, maxb).}
\usage{
boundp(x, minb, maxb, hbf = 0)

boundp(x, minb, maxb, hbf = 0)
}
\arguments{
\item{x}{The value to transform.}

\item{minb}{The minimum bound.}

\item{maxb}{The maximum bound.}

\item{hbf}{The hybrid bounded flag. Determines which function is used.}

\item{x}{The value to transform.}

\item{minb}{The minimum bound.}

\item{maxb}{The maximum bound.}

\item{hbf}{The hybrid bounded flag. Determines which function is used.}
}
\description{
If \code{hbf=0} then the function is
minb+(maxb-minb)/(1+exp(-x)), otherwise it is
minb+(maxb-minb)*(.5*sin(x*pi/2)+.5).

If \code{hbf=0} then the function is
minb+(maxb-minb)/(1+exp(-x)), otherwise it is
minb+(maxb-minb)*(.5*sin(x*pi/2)+.5).
}

