% Generated by roxygen2 (4.0.1): do not edit by hand
\name{boundpin}
\alias{boundpin}
\title{Inverse bounding transformation function used by ADMB.}
\usage{
boundpin(x, minb, maxb, hbf)
}
\arguments{
\item{x}{The value to back-transform.}

\item{minb}{The minimum bound.}

\item{maxb}{The maximum bound.}

\item{hbf}{The hybrid bounded flag. Determines which function is used.}
}
\description{
Inverse bounding transformation function used by ADMB.
}

