% Generated by roxygen2 (4.0.0): do not edit by hand
\name{get.admb.cov}
\alias{get.admb.cov}
\title{Read in an ADMB admodel.cov file}
\usage{
get.admb.cov(model.path = getwd())

get.admb.cov(model.path = getwd())
}
\arguments{
\item{model.path}{The directory path to the folder containing the model.}

\item{model.path}{The directory path to the folder containing the model.}
}
\value{
A list of the number of parameters (\code{num.pars}), the
covariance matrix in unbounded space (\code{cov}), the hybrid bounded
flag (\code{hybrid_bounded_flat}), and the \code{scales}, which are the
second derivatives of the bounding functions evaluated at the MLEs.

A list of the number of parameters (\code{num.pars}), the
covariance matrix in unbounded space (\code{cov}), the hybrid bounded
flag (\code{hybrid_bounded_flat}), and the \code{scales}, which are the
second derivatives of the bounding functions evaluated at the MLEs.
}
\description{
This function reads in all of the information contained in the
admodel.cov file. ed at the MLE (used to
convert from unbounded to bounded space).

This function reads in all of the information contained in the
admodel.cov file. ed at the MLE (used to
convert from unbounded to bounded space).
}
\details{
The admodel.cov file is written by ADMB after the Hessian
matrix is inverted.

The admodel.cov file is written by ADMB after the Hessian
matrix is inverted.
}
\author{
Cole Monnahan

Cole Monnahan
}

