% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get.admb.fit}
\alias{get.admb.fit}
\title{Read in the maximum likelihood results of an ADMB model}
\usage{
get.admb.fit(model.name, model.path = getwd())

get.admb.fit(model.name, model.path = getwd())
}
\arguments{
\item{model.name}{The name of the model (without the .tpl extension).}

\item{model.path}{The directory path to the folder containing the model.}

\item{model.name}{The name of the model (without the .tpl extension).}

\item{model.path}{The directory path to the folder containing the model.}
}
\description{
Read in the maximum likelihood results of an ADMB model

Read in the maximum likelihood results of an ADMB model
}
\author{
Cole Monnahan, modified from a version found at
http://qfc.fw.msu.edu/userfun.asp

Cole Monnahan, modified from a version found at
http://qfc.fw.msu.edu/userfun.asp
}

