% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ndfboundp}
\alias{ndfboundp}
\title{Inverse bounding transformation function used by ADMB.}
\usage{
ndfboundp(x, minb, maxb, hbf)
}
\arguments{
\item{x}{The value for which}

\item{minb}{The minimum bound.}

\item{maxb}{The maximum bound.}

\item{hbf}{The hybrid bounded flag. Determines which function is used.}
}
\description{
Inverse bounding transformation function used by ADMB.
}

