% Generated by roxygen2 (4.0.1): do not edit by hand
\name{run.mcmc}
\alias{run.mcmc}
\title{Run an MCMC using an ADMB model, return (1) the posterior draws, MLE
fits and covariance/correlation matrices, and some MCMC convergence
diagnostics using CODA.}
\usage{
run.mcmc(model.path, model.name, Nout, mcsave, burn.in, cor.mat = NULL,
  init.pin = NULL, se.scale = NULL, mcscale = FALSE, mcseed = NULL,
  mcrb = NULL, mcprobe = NULL, verbose = TRUE, extra.args = NULL)
}
\arguments{
\item{model.path}{A path to the folder containing the model.}

\item{mode.name}{The name of the model executable. A character string,
without '.exe'.}

\item{Nout}{The number of draws after thinning and burn in.}

\item{mcsave}{Controls thinning of samples. Save every mcsave value,
such that 1 corresponds to keeping all draws, and 100 saving every 100th
draw.}

\item{burn.in}{How many samples to discard from the beginning of the
chain, *after* thining. The burn in period (i.e., the first
burn.in*mcsave draws) should be at least large enough to cover dynamic
scaling.}

\item{cor.mat}{A manually defined correlation matrix to use in the
Metropolis-Hastings algorithm.}

\item{init.pin}{A vector of initial values, which are written to file
and used in the model via the -mcpin option.}

\item{se.scale}{A value which scales all of the variances from the MLE
fit. A value of 1 indicates to use the estimated variances.}

\item{mcscale}{Whether to use the mcscale option, which dynamically
scales the covariance matrix for efficient acceptance ratios.}

\item{mcseed}{Which seed (integer value) to pass ADMB. Used for
reproducibility.}

\item{mcrb}{Which value to use in the rescale bounded algorithm. Must be
an integer from 1-9. The default NULL value disables this feature. See
the vignette for more information on this algorithm and how to best use
it.}

\item{mcprobe}{Which value to use in the probing algorithm. The default
NULL value disables this feature. See the vignette for more information
on this algorithm and how to best use it.}

\item{verbose}{Boolean value for whether to print ADMB warnings and other
information. Useful for verifying the code is interacting with a
properly functioning model.}

\item{extra.args}{A character string which is passed to ADMB at
runtime. Useful for passing additional arguments to the model
executable.}
}
\description{
Run an MCMC using an ADMB model, return (1) the posterior draws, MLE
fits and covariance/correlation matrices, and some MCMC convergence
diagnostics using CODA.
}

