% Generated by roxygen2 (4.0.1): do not edit by hand
\name{run_admb_mcmc}
\alias{run_admb_mcmc}
\title{Run an MCMC using an ADMB model, return (1) the posterior draws, MLE
fits and covariance/correlation matrices, and some MCMC convergence
diagnostics using CODA.}
\usage{
run_admb_mcmc(model.path, model.name, Nout, mcsave, burn.in, cov.user = NULL,
  init.pin = NULL, se.scale = NULL, mcscale = FALSE, mcseed = NULL,
  mcrb = NULL, mcdiag = FALSE, mcprobe = NULL, verbose = TRUE,
  extra.args = NULL, hybrid = FALSE, hyeps = NULL, hynstep = NULL)
}
\arguments{
\item{model.path}{(Character) A path to the folder containing the model. NULL
indicates the current folder.}

\item{mode.name}{(Character) The name of the model executable. A character string,
without '.exe'.}

\item{Nout}{(Integer) The number of draws after thinning and burn in.}

\item{mcsave}{(Integer) Controls thinning of samples. Save every mcsave
value, such that 1 corresponds to keeping all draws, and 100 saving
every 100th draw.}

\item{burn.in}{(Integer) How many samples to discard from the beginning
of the chain, *after* thining. The burn in period (i.e., the first
burn.in*mcsave draws) should be at least large enough to cover dynamic
scaling.}

\item{cov.user}{(Numeric matrix) A manually defined covariance matrix (in bounded space)
to use in the Metropolis-Hastings algorithm.}

\item{init.pin}{(Numeric vector) A vector of initial values, which are written to file
and used in the model via the -mcpin option.}

\item{se.scale}{(Numeric) A value which scales all of the variances from
the MLE fit. A value of 1 indicates to use the estimated variances.}

\item{mcscale}{(Logical) Whether to use the mcscale option, which
dynamically scales the covariance matrix for efficient acceptance
ratios.}

\item{mcseed}{(Integer) Which seed (integer value) to pass ADMB. Used
for reproducibility.}

\item{mcrb}{(Integer) Which value to use in the rescale bounded
algorithm. Must be an integer from 1-9. The default NULL value disables
this feature. See the vignette for more information on this algorithm
and how to best use it.}

\item{mcdiag}{(Logical) Whether to use the \code{mcdiag} feature. This
uses an identity matrix for the covariance matrix.  #'}

\item{mcprobe}{Which value to use in the probing algorithm. The default NULL value
disables this feature. See the vignette for more information on this
algorithm and how to best use it.}

\item{verbose}{(Logical) Whether to print ADMB warnings and other
information. Useful for testing and troubleshooting.}

\item{extra.args}{(Character) A string which is passed to ADMB at
runtime. Useful for passing additional arguments to the model
executable.}
}
\value{
Returns a list containing (1) the posterior draws, (2) and
object of class 'admb', read in using the results read in using
\code{read_admb}, and (3) some MCMC convergence diagnostics using CODA.
}
\description{
Run an MCMC using an ADMB model, return (1) the posterior draws, MLE
fits and covariance/correlation matrices, and some MCMC convergence
diagnostics using CODA.
}

