% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_ss_mcmc.R
\name{run_ss_mcmc_parallel}
\alias{run_ss_mcmc_parallel}
\title{A wrapper for running SS models in parallel}
\usage{
run_ss_mcmc_parallel(parallel_number, Nout, mcsave, burn.in, cov.user = NULL,
  extra.args = NULL, seed, se.scale = 1)
}
\description{
A wrapper for running SS models in parallel
}

