% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounding_functions.R
\name{boundp}
\alias{boundp}
\title{The bounding function.}
\usage{
boundp(x, minb, maxb, hbf = 0)
}
\arguments{
\item{x}{The input value.}

\item{minb}{The minimum bound.}

\item{maxb}{The maximum bound.}

\item{hbf}{The hybrid bounded flag. Determines which function is used.}
}
\value{
Numeric function value.
}
\description{
The bounding function.

The bounding functions, with their inverses and derivatives.
}
\details{
If \code{hbf=0} then the function is minb+(maxb-minb)/(1+exp(-x)),
otherwise it is minb+(maxb-minb)*(.5*sin(x*pi/2)+.5).
}
\section{Note}{
 These functions are named as in the ADMB source code.
}
\author{
Cole C. Monnahan
}

