% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dat_file.R
\name{create_dat_file}
\alias{create_dat_file}
\title{Create .dat file for use in ADMB model}
\usage{
create_dat_file(dat_list, dat_name, d = getwd(), f = "ADMB")
}
\arguments{
\item{dat_list}{list with data that will be written to .dat file}

\item{d}{directory where .dat file will be placed. Default is the working directory.}

\item{f}{file name for .dat file. Default is ADMB.dat}
}
\description{
\code{create_dat_file} takes a list of data \code{dat_list},
}
\details{
\code{dat_list} must be a \code{list}, if not an error is thrown.
}
