% Generated by roxygen2 (4.0.0): do not edit by hand
\name{get.admb.hes}
\alias{get.admb.hes}
\title{Read in an ADMB admodel.hes file}
\usage{
get.admb.hes(model.path = getwd())

get.admb.hes(model.path = getwd())
}
\arguments{
\item{model.path}{The directory path to the folder containing the model.}

\item{model.path}{The directory path to the folder containing the model.}
}
\value{
A list of the number of parameters (\code{num.pars}), the
Hessian matrix in unbounded space (\code{hes}), the hybrid bounded
flag (\code{hybrid_bounded_flat}), and the \code{scales}, which are the
second derivatives of the bounding functions evaluated at the MLEs.

A list of the number of parameters (\code{num.pars}), the
Hessian matrix in unbounded space (\code{hes}), the hybrid bounded
flag (\code{hybrid_bounded_flat}), and the \code{scales}, which are the
second derivatives of the bounding functions evaluated at the MLEs.
}
\description{
This function reads in all of the information contained in the
admodel.hes file. It returns a list w/ the number of parameters, the
Hessian matrix (in unbounded space), the hybrid bounded flag (determines
the bounding function), and the scales, which are the second derivatives
of the bounding functions evaluated at the MLE (used to convert from
unbounded to bounded space).

This function reads in all of the information contained in the
admodel.hes file. It returns a list w/ the number of parameters, the
Hessian matrix (in unbounded space), the hybrid bounded flag (determines
the bounding function), and the scales, which are the second derivatives
of the bounding functions evaluated at the MLE (used to convert from
unbounded to bounded space).
}
\details{
The admodel.hes file is written by ADMB during the optimization
phase. The Hessian matrix contains the second derivatives evaluated at
the maximum likelihood estimates. The inverse of the Hessian is the
covariance matrix used to determine standard errors and in the MCMC
algorithm. This file is written even if the matrix is not invertible
(not positive definite).

The admodel.hes file is written by ADMB during the optimization
phase. The Hessian matrix contains the second derivatives evaluated at
the maximum likelihood estimates. The inverse of the Hessian is the
covariance matrix used to determine standard errors and in the MCMC
algorithm. This file is written even if the matrix is not invertible
(not positive definite).
}
\author{
Cole Monnahan

Cole Monnahan
}

